+++ CHANGELOG.md
@@ -5,6 +5,7 @@ https://keepachangelog.com/en/1.0.0/
 - Fixed: `ElementaryMethodsOfTypeInCollectionNotOverridden` triggers for external types
 - Fixed: `ExceptionThrownFromProhibitedContext` will no longer trigger for `NotSupportedException` and `NotImplementedException`
 - Fixed: `TestMethodWithoutTestAttribute` no longer crashes when encountering a `record`
+- Fixed: `TestMethodWithoutTestAttribute` no longer triggers for `Dispose()` methods
 
 ## [1.2.0] - 2022-08-28
 - Implemented `UnnecessaryEnumerableMaterialization`: An IEnumerable was materialized before a deferred execution call